Defocus blur and dielectric material

+++ src/hitable.h
@@ -19,7 +19,7 @@ struct hit_record
 class hitable
 {
 public:
-  virtual bool hit(const ray &r, float t_min, float t_max, hit_record &rec) const = 0;
+  virtual bool hit(const ray &r, double t_min, double t_max, hit_record &rec) const = 0;
 };
 
 #endif // !HITABLE_H
