+++ src/hitable_list.h
@@ -8,13 +8,13 @@ class hitable_list : public hitable
 public:
   hitable_list() {};
   hitable_list(hitable **l, int n) : list(l), list_size(n) {};
-  bool hit (const ray &r, float t_min, float t_max, hit_record &rec) const override;
+  bool hit (const ray &r, double t_min, double t_max, hit_record &rec) const override;
 
   hitable **list;
   int list_size;
 };
 
-bool hitable_list::hit (const ray &r, float t_min, float t_max, hit_record &rec) const
+bool hitable_list::hit (const ray &r, double t_min, double t_max, hit_record &rec) const
 {
   hit_record temp_rec;
   bool hit_anything = false;
