+++ src/sphere.h
@@ -3,19 +3,19 @@
 
 #include "hitable.h"
 
-class sphere : public hitable 
+class sphere : public hitable
 {
 public:
   sphere() : center(), radius(0) {}
   sphere(vec3 pos, double r, material *m) : center(pos), radius(r), mat(m) {}
-  bool hit(const ray &r, float t_min, float t_max, hit_record &rec) const override;
-  
+  bool hit(const ray &r, double t_min, double t_max, hit_record &rec) const override;
+
   vec3 center;
   double radius;
   material *mat;
 };
 
-bool sphere::hit(const ray &r, float t_min, float t_max, hit_record &rec) const
+bool sphere::hit(const ray &r, double t_min, double t_max, hit_record &rec) const
 {
   // sphere on p(x, y, z):
   // x*x + y*y + z*z = r*r
@@ -30,7 +30,7 @@ bool sphere::hit(const ray &r, float t_min, float t_max, hit_record &rec) const
   // dot((A + t*B)-C, (A + t*B)-C) = r*r
   // ((A + t*B)-C)*((A + t*B)-C) = r*r
   // discriminant on ax^2+bx+c=0 is b^2-4ac
-  // if discrim is 0 has one lolution, 
+  // if discrim is 0 has one lolution,
   // if < 0 has no solition and if > 0 has 2 solutions
 
   vec3 oc = r.origin() - center;
